backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
motor>=3.3.2
pydantic>=2.6.1
pydantic-settings>=2.2.1
dnspython>=2.6.0
python-dotenv>=1.0.1
